Genetic Network
Adds the Network class representing one single network capable of the forward propagation. The backprop depends on multiple networks, so it needs to be separated. The cycle is handled by the NetworkHandler class, which calls multiple forwardprops from different nets and then calls the GeneticLearning class with the net outputs